requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.12.5
cffi
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
joblib==1.0.1
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.36.0
matplotlib==3.4.1
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
moviepy==1.0.3
numba==0.53.1
numpy
olefile==0.46
opencv-python==4.5.1.48
packaging==20.9
Pillow
pooch==1.3.0
prefetch-generator==1.0.1
proglog==0.1.9
pycparser
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
resampy==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
six
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.60.0
urllib3==1.26.4
wincertstore==0.2
